Bio::Tools::Phylo::PAML: fix reading of results data.
The parsing of the summary was actually reading the whole file.  This
issue was not unique to yn00 output and to reading of its ML matrix.
See the failing test where it also failed for some calls of codeml.
So generalize the fix and rewind file to the start after parsing
summary.